requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
base58==2.1.1
bitarray==3.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
coinbase-advanced-py==1.8.2
construct==2.10.70
construct-typing==0.7.0
cryptography==46.0.3
cytoolz==1.1.0
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonalias==0.1.1
multidict==6.7.0
nest-asyncio==1.6.0
parsimonious==0.10.0
propcache==0.4.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0
python-dotenv==1.2.1
pyunormalize==17.0.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
six==1.17.0
solana==0.36.10
solders==0.27.1
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
web3==7.10.0
websockets==13.1
yarl==1.22.0
